+++ app/src/main/java/com/teambeme/beme/data/remote/datasource/HomeDataSource.kt
@@ -4,6 +4,7 @@ import com.teambeme.beme.presentation.home.model.RequestModifyPublic
 import com.teambeme.beme.presentation.home.model.ResponseAnswer
 import com.teambeme.beme.presentation.home.model.ResponseAnswers
 import com.teambeme.beme.presentation.home.model.ResponseModifyData
+import retrofit2.Call
 
 interface HomeDataSource {
     suspend fun modifyPublic(body: RequestModifyPublic): ResponseModifyData
@@ -11,4 +12,5 @@ interface HomeDataSource {
     suspend fun getNewAnswer(): ResponseAnswer
     suspend fun changeQuestion(answerId: Int): ResponseAnswer
     suspend fun deleteAnswer(answerId: Int): ResponseModifyData
+    suspend fun fetchAnswerPagingData(page: Int): Call<ResponseAnswers>
 }
