+++ app/src/main/java/com/teambeme/beme/data/repository/HomeRepositoryImpl.kt
@@ -1,11 +1,17 @@
 package com.teambeme.beme.data.repository
 
+import androidx.paging.Pager
+import androidx.paging.PagingConfig
+import com.teambeme.beme.base.BeMePagingSource
+import com.teambeme.beme.base.NETWORK_PAGE_SIZE
+import com.teambeme.beme.data.remote.api.HomeService
 import com.teambeme.beme.data.remote.datasource.HomeDataSource
 import com.teambeme.beme.domain.repository.HomeRepository
 import com.teambeme.beme.presentation.home.model.RequestModifyPublic
 import com.teambeme.beme.presentation.home.model.ResponseAnswer
 import com.teambeme.beme.presentation.home.model.ResponseAnswers
 import com.teambeme.beme.presentation.home.model.ResponseModifyData
+import retrofit2.await
 import javax.inject.Inject
 
 class HomeRepositoryImpl @Inject constructor(
@@ -24,4 +30,16 @@ class HomeRepositoryImpl @Inject constructor(
 
     override suspend fun deleteAnswer(answerId: Int): ResponseModifyData =
         homeDataSource.deleteAnswer(answerId)
+
+    override fun retrieveAnswerPages() = Pager(
+        config = PagingConfig(NETWORK_PAGE_SIZE),
+        pagingSourceFactory = {
+            BeMePagingSource {
+                homeDataSource.fetchAnswerPagingData(it)
+                    .await()
+                    .answers
+            }
+        }
+    ).flow
+
 }
