+++ app/src/main/java/com/teambeme/beme/presentation/setting/model/OpenSourceData.kt
@@ -1,4 +1,4 @@
-package com.teambeme.beme.setting.model
+package com.teambeme.beme.presentation.setting.model
 
 data class OpenSourceData(
     val openSource: String,
