+++ app/src/main/res/layout/activity_open_source.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".setting.view.OpenSourceActivity">
+        tools:context=".presentation.setting.view.OpenSourceActivity">
 
         <androidx.appcompat.widget.Toolbar
             android:id="@+id/toolbar_opensource"
